:hammer: Refactor: 컨벤션 수정 및 컴포넌트 분리
