+++ src/components/ChatRoom/ChatBubble.tsx
@@ -1,17 +1,10 @@
-import { styled } from 'styled-components';
-import { messageData } from '@/apis/dummy_ChatAPI';
-import { isSenderMe } from '@/components/ChatRoom/ChatFunctions';
 import { useEffect, useRef } from 'react';
-import ChatAlert from '@/components/ChatRoom/ChatAlert';
-import ChatProfile from '@/components/ChatRoom/ChatProfile';
+import { styled } from 'styled-components';
 
-type TMessageContent = {
-  messageId: number;
-  messageType: string;
-  content?: string;
-  senderId: number;
-  sendTime: string;
-};
+import { messageData } from '@/apis/dummy_ChatAPI';
+import ChatAlert from '@/components/ChatRoom/ChatAlert';
+import ChatBubbleRender from '@/components/ChatRoom/ChatBubbleRender';
+
 const ChatBubble = () => {
   const messageEndRef = useRef<HTMLDivElement | null>(null);
   //맨 끝으로 렌더
@@ -19,39 +12,15 @@ const ChatBubble = () => {
     messageEndRef.current?.scrollIntoView({ behavior: 'auto' });
   }, []);
 
-  //말풍선 렌더
-  const renderChatStyledBubbles = (messageContent: TMessageContent, index: number) => {
-    //이전의 메세지와 보내는 사람이 같은지 판별
-    const isSameAuthorAsPrevious =
-      index > 0 && messageData.messages[index - 1].senderId === messageContent.senderId;
-    //보내는 사람이 나와 같은지 판별
-    const isAuthorMe = isSenderMe(messageContent.senderId);
-    //이전의 메세지가 같은 타입인지 판별
-    const isTypeSameAsPrevious =
-      index > 0 && messageData.messages[index - 1].messageType === messageContent.messageType;
-    //프로필이 필요한지 판별
-    const isProfileNeed =
-      (!isAuthorMe && !isSameAuthorAsPrevious) || (!isAuthorMe && !isTypeSameAsPrevious);
-
-    const gapSize = isSameAuthorAsPrevious && isTypeSameAsPrevious ? '6px' : '16px';
-
-    return (
-      <StyledStyledWrap
-        key={messageContent.messageId}
-        style={{ marginTop: gapSize }}>
-        {isProfileNeed ? <ChatProfile senderId={messageContent.senderId} /> : null}
-        <StyledBubbles className={isAuthorMe ? 'myBubbles' : 'othersBubbles'}>
-          <p>{messageContent.content}</p>
-        </StyledBubbles>
-      </StyledStyledWrap>
-    );
-  };
-
   return (
     <StyledContainer>
       {messageData.messages.map((messageContent, index) => {
         return messageContent.messageType === 'TALK' ? (
-          renderChatStyledBubbles(messageContent, index)
+          <ChatBubbleRender
+            messageContent={messageContent}
+            index={index}
+            key={messageContent.messageId}
+          />
         ) : (
           <ChatAlert
             senderId={messageContent.senderId}
@@ -75,30 +44,6 @@ const StyledContainer = styled.div`
   scrollbar-width: none; /* Firefox */
   &::-webkit-scrollbar {
     display: none; /* Chrome, Safari, Opera*/
-  }
-`;
-const StyledStyledWrap = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-const StyledBubbles = styled.div`
-  padding: 11px 16px;
-  border-radius: 20px;
-  font-size: 14px;
-  overflow-wrap: break-word;
-  word-break: break-word;
-  &.myBubbles {
-    max-width: calc(100% - 80px);
-    color: ${({ theme }) => theme.colors.white};
-    background: ${({ theme }) => theme.colors.ctaColor};
-    text-align: right;
-    margin-left: auto;
-  }
-  &.othersBubbles {
-    max-width: calc(100% - 120px);
-    margin-left: 60px;
-    margin-right: auto;
-    background: ${({ theme }) => theme.colors.grayColor1};
   }
 `;
