+++ src/components/ChatRoom/ChatDeclarationStep3.tsx
@@ -1,5 +1,6 @@
 import styled from 'styled-components';
 import { useSetRecoilState } from 'recoil';
+
 import {
   declarationStepAtom,
   isDeclarationBottomsheetOpenAtom
