+++ src/components/ChatRoom/ChatInput.tsx
@@ -1,5 +1,6 @@
-import { styled } from 'styled-components';
 import { useState } from 'react';
+import { styled } from 'styled-components';
+
 import IconSend from '@/assets/chatRooms/ico_send.svg';
 
 const ChatInput = () => {
