+++ src/constants/declarationCategory.ts
@@ -1,0 +1,11 @@
+export const DECLARATION_DATA = [
+  '홍보/상업성',
+  '같은 내용 도배',
+  '욕설/인신공격',
+  '음란/선정성',
+  '불법정보',
+  '개인정보 노출',
+  '권리침해 신고',
+  '관련 없는 내용',
+  '기타'
+];
